src/qa_agent/requirements.txt
streamlit
typing_extensions
langgraph
langchain_groq
tavily-python
groq
langchain_core
